Finally implements the fast-multiplication optimization from https://github.com/JuliaMath/FixedPointDecimals.jl/pull/45, but this time for 128-bit FixedDecimals! :)

This is a follow-up to
https://github.com/JuliaMath/FixedPointDecimals.jl/pull/93, which introduces an Int256 type for widemul. However, the fldmod still required 2 BigInt allocations.

Now, this PR uses a custom implementation of the LLVM div-by-const optimization for (U)Int256, which briefly widens to Int512 (😅) to perform the fldmod by the constant 10^f coefficient.

This brings 128-bit FD multiply to the same performance as 64-bit. :)

master:
```julia
julia> @btime fd * fd setup = (fd = FixedDecimal{Int32,3}(1.234))
  0.833 ns (0 allocations: 0 bytes)
FixedDecimal{Int32,3}(1.523)

julia> @btime fd * fd setup = (fd = FixedDecimal{Int64,3}(1.234))
  0.833 ns (0 allocations: 0 bytes)
FixedDecimal{Int64,3}(1.523)

julia> @btime fd * fd setup = (fd = FixedDecimal{Int128,3}(1.234))
  398.960 ns (24 allocations: 464 bytes)
FixedDecimal{Int128,3}(1.523)

julia> @btime fd * fd setup = (fd = FixedDecimal{UInt128,3}(1.234))
  455.357 ns (30 allocations: 592 bytes)
FixedDecimal{UInt128,3}(1.523)
```
After PR #93:
```julia
julia> @btime fd * fd setup = (fd = FixedDecimal{Int32,3}(1.234))
  0.792 ns (0 allocations: 0 bytes)
FixedDecimal{Int32,3}(1.523)

julia> @btime fd * fd setup = (fd = FixedDecimal{Int64,3}(1.234))
  0.833 ns (0 allocations: 0 bytes)
FixedDecimal{Int64,3}(1.523)

julia> @btime fd * fd setup = (fd = FixedDecimal{Int128,3}(1.234))
  206.591 ns (12 allocations: 240 bytes)
FixedDecimal{Int128,3}(1.523)

julia> @btime fd * fd setup = (fd = FixedDecimal{UInt128,3}(1.234))
  264.194 ns (16 allocations: 432 bytes)
FixedDecimal{UInt128,3}(1.523)
```
After this PR:
```julia
julia> @btime fd * fd setup = (fd = FixedDecimal{Int32,3}(1.234))
  0.833 ns (0 allocations: 0 bytes)
FixedDecimal{Int32,3}(1.523)

julia> @btime fd * fd setup = (fd = FixedDecimal{Int64,3}(1.234))
  0.792 ns (0 allocations: 0 bytes)
FixedDecimal{Int64,3}(1.523)

julia> @btime fd * fd setup = (fd = FixedDecimal{Int128,3}(1.234))
  0.791 ns (0 allocations: 0 bytes)
FixedDecimal{Int128,3}(1.523)

julia> @btime fd * fd setup = (fd = FixedDecimal{UInt128,3}(1.234))
  0.792 ns (0 allocations: 0 bytes)
FixedDecimal{UInt128,3}(1.523)
```