@NHDaly 2024-06-13
Huzzah! I finally feel like we can bring the ideas from https://github.com/JuliaMath/FixedPointDecimals.jl/pull/45 to this package, and it's still valuable. It turns out that LLVM has already applied this optimization automatically for Int128 div by const (thus * for `FixedDecimal{Int64}` is already fast). But this PR adds support for `FixedDecimal{Int128}`, which needed a custom function for Int256 div by const. :)

See here, for my note on FixedDecimal{Int64} support:
https://github.com/JuliaMath/FixedPointDecimals.jl/pull/45#issuecomment-2163836966