@NHDaly 2024-06-13
Now that there's julia's effects system, this doesn't need `@pure`, so it's actually possible to do this! 
And now that we've introduced BitIntegers.jl, the code is way simpler. 😊 woohoo!