@NHDaly 2024-06-13
Update 1: It turned out that (of course) `*` is too fast for `@btime` to get a good measurement of on its own. I've adjusted the benchmark to perform repeated multiplies, and the true performance is now included in the original comment.

Update 2: It turns out that this is actually faster for _all_ int types (>2x faster for FD{Int64}), even though LLVM was doing a version of this optimization on its own. I'm honestly pretty surprised by that... From what I can tell, for FD{Int32}, the code is _identical_ except for a single instruction changes from comparing `hi` to comparing `gt`:
<img width="1503" alt="Screenshot 2024-06-13 at 2 10 20 PM" src="https://github.com/JuliaMath/FixedPointDecimals.jl/assets/1582097/df9715e0-baff-4ca7-8217-33b5a11c72e4">
And yet somehow it's 25% faster (80µs -> 60µs).