@NHDaly 2024-06-13
Aha, the native code difference comes from the whole benchmark function. Before, the multiply was outlined, and called through a function call.
Now, it's inlined, and i think the whole loop is constant-folded away, so it's actually sort of cheating. 😅 
But still, allowing that optimization seems like a positive thing!