@EdsterG 2024-09-16
By the way, `div`/`mod` allocations are fixed in the most recent version of BitIntegers when running on 1.11+, see [#48](https://github.com/rfourquet/BitIntegers.jl/pull/48).