@NHDaly 2024-09-26
Oh wow, very nice! Thanks @EdsterG. 💪 

This should still help quite a bit, but the baseline for comparison would be better in 1.11, then! 💪